requirements.txt
Flask
python-dotenv
python-telegram-bot==13.15